Feat/channel auto create cfg

See merge request !19

+++ nsqdserver/http.go
@@ -611,6 +611,12 @@ func (s *httpServer) doCreateChannel(w http.ResponseWriter, req *http.Request, p
 		return nil, err
 	}
 	if s.ctx.checkConsumeForMasterWrite(topic.GetTopicName(), topic.GetTopicPart()) {
+		//check channel exists in topic partition info, if topic has DisableChannelAutoCreate
+		dynamicInfo := topic.GetDynamicInfo()
+		if dynamicInfo.DisableChannelAutoCreate {
+			return nil, http_api.Err{403, fmt.Sprintf("channel %v not registered. channel should be initialized before use.", channelName)}
+		}
+
 		topic.GetChannel(channelName)
 		// need sync channel after created
 		err := s.ctx.SyncChannels(topic)
@@ -763,6 +769,10 @@ func (s *httpServer) doDeleteChannel(w http.ResponseWriter, req *http.Request, p
 	}
 
 	if s.ctx.checkConsumeForMasterWrite(topic.GetTopicName(), topic.GetTopicPart()) {
+		dynamicInfo := topic.GetDynamicInfo()
+		if dynamicInfo.DisableChannelAutoCreate {
+			return nil, http_api.Err{403, fmt.Sprintf("topic %v has channel auto create disabled. channel %v cuold not be deleted localy in nsqd.", topic.GetTopicName(), channelName)}
+		}
 		clusterErr := s.ctx.DeleteExistingChannel(topic, channelName)
 		if clusterErr != nil {
 			return nil, http_api.Err{500, clusterErr.Error()}
